// forgeos-platform/sdk/packages/core/src/forgeos-core/types.ts
export interface ForgeOSCoreStartInput
	extends Omit<StartSessionInput, "config" | "localRuntime"> {
	config: ForgeOSCoreStartConfig;
	localRuntime?: LocalRuntimeStartOptions;
}